Updated install to better detect OS. Also auto run markdown setup script

+++ install.sh
@@ -4,33 +4,41 @@ INSTALL_DIR="$HOME/.config/nvim"
 GIT_SOURCE="https://github.com/jgrim/nvim.git"
 
 # Install system packages
-BREW=$(which brew)
-APT=$(which apt)
-if [[ ! -z "$BREW" ]]; then
-    $BREW cask install adoptopenjdk --no-quarantine
-    $BREW install git nvm neovim ctags global composer python python3 php@7.4 ripgrep languagetool
-fi
-if [[ ! -z "$APT" ]]; then
-    PHP_MODULES="php7.4 php7.4-cli php7.4-common php7.4-curl php7.4-curl php7.4-gd php7.4-intl php7.4-json php7.4-mbstring php7.4-opcache php7.4-readline php7.4-soap php7.4-xml"
-    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
-    sudo $APT install git neovim ctags global python3 python3-pip $PHP_MODULES ripgrep 
-
-    # Install composer if it's not installed
-    if ! type "composer" > /dev/null; then
-        php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-        php -r "if (hash_file('sha384', 'composer-setup.php') === '756890a4488ce9024fc62c56153228907f1545c228516cbf63f885e036d37e9a59d27d63f46af1d4d07ee0f76181c7d3') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
-        sudo php composer-setup.php --install-dir=/usr/local/bin --filename=composer
-        php -r "unlink('composer-setup.php');"
+if [ "$(uname)" == "Darwin" ]; then
+    BREW=$(which brew)
+    if [[ ! -z "$BREW" ]]; then
+        $BREW install --cask adoptopenjdk --no-quarantine
+        $BREW install git nvm neovim ctags global composer python python3 php@7.4 ripgrep languagetool
+    else
+        echo "Please install Homebrew"
+        exit 2
     fi
-    composer selfupdate > /dev/null
-    sudo snap install languagetool &> /dev/null
+elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
+    APT=$(which apt)
+    if [[ ! -z "$APT" ]]; then
+        PHP_MODULES="php7.4 php7.4-cli php7.4-common php7.4-curl php7.4-curl php7.4-gd php7.4-intl php7.4-json php7.4-mbstring php7.4-opcache php7.4-readline php7.4-soap php7.4-xml"
+        curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
+        sudo $APT install git neovim ctags global python3 python3-pip $PHP_MODULES ripgrep 
+
+        # Install composer if it's not installed
+        if ! type "composer" > /dev/null; then
+            php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+            php -r "if (hash_file('sha384', 'composer-setup.php') === '756890a4488ce9024fc62c56153228907f1545c228516cbf63f885e036d37e9a59d27d63f46af1d4d07ee0f76181c7d3') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
+            sudo php composer-setup.php --install-dir=/usr/local/bin --filename=composer
+            php -r "unlink('composer-setup.php');"
+        fi
+        composer selfupdate > /dev/null
+        sudo snap install languagetool &> /dev/null
+    fi
+else
+    echo "Unsupported OS"
+    exit 2
 fi
 
 # Install latest node
 NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
 [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-
 nvm install stable && npm install -g neovim
 
 # Install python for neovim
@@ -83,4 +91,4 @@ composer global require friendsofphp/php-cs-fixer
 composer global require "squizlabs/php_codesniffer=*"
 
 nvim +'PlugInstall --sync' +qa
-
+nvim +'mkdp#util#install()' +qa
